requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==41.0.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.108.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.13.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
sqlmodel==0.0.14
starlette==0.32.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
